un7z_semantic3d.sh: Check if '*.labels' already moved

+++ data_conversions/un7z_semantic3d.sh
@@ -31,10 +31,10 @@ unpack $BASE_DIR/train/sg28_station4_intensity_rgb.7z
 unpack $BASE_DIR/train/untermaederbrunnen_station1_xyz_intensity_rgb.7z
 unpack $BASE_DIR/train/sem8_labels_training.7z
 
-mv $BASE_DIR/train/bildstein_station3_xyz_intensity_rgb.labels $BASE_DIR/val
-mv $BASE_DIR/train/domfountain_station2_xyz_intensity_rgb.labels $BASE_DIR/val
-mv $BASE_DIR/train/sg27_station4_intensity_rgb.labels $BASE_DIR/val
-mv $BASE_DIR/train/untermaederbrunnen_station3_xyz_intensity_rgb.labels $BASE_DIR/val
+[ -f $BASE_DIR/val/bildstein_station3_xyz_intensity_rgb.labels ] || mv $BASE_DIR/train/bildstein_station3_xyz_intensity_rgb.labels $BASE_DIR/val
+[ -f $BASE_DIR/val/domfountain_station2_xyz_intensity_rgb.labels ] || mv $BASE_DIR/train/domfountain_station2_xyz_intensity_rgb.labels $BASE_DIR/val
+[ -f $BASE_DIR/val/sg27_station4_intensity_rgb.labels ] || mv $BASE_DIR/train/sg27_station4_intensity_rgb.labels $BASE_DIR/val
+[ -f $BASE_DIR/val/untermaederbrunnen_station3_xyz_intensity_rgb.labels ] || mv $BASE_DIR/train/untermaederbrunnen_station3_xyz_intensity_rgb.labels $BASE_DIR/val
 
 # Validation data
 unpack $BASE_DIR/val/bildstein_station3_xyz_intensity_rgb.7z
